# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger API for rev tracker.",
        "title": "RevTracker server Swagger API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "erupshis",
            "email": "e.rupshis@gmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/data": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Insert godoc",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "homework_data"
                ],
                "summary": "adds new homework data",
                "operationId": "data-insert",
                "parameters": [
                    {
                        "description": "data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.Data"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.Data"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/data/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Select godoc",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "homework_data"
                ],
                "summary": "select new data",
                "operationId": "data-select",
                "parameters": [
                    {
                        "type": "string",
                        "description": "homework id",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.Data"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update godoc",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "homework_data"
                ],
                "summary": "updates homework data",
                "operationId": "data-update",
                "parameters": [
                    {
                        "description": "updated homework data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.Data"
                        }
                    },
                    {
                        "type": "string",
                        "description": "homework id",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.Data"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete godoc",
                "tags": [
                    "homework_data"
                ],
                "summary": "deletes existing data",
                "operationId": "data-delete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "homework id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/homework/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Select godoc",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "homework"
                ],
                "summary": "select homework",
                "operationId": "homework-select",
                "parameters": [
                    {
                        "type": "string",
                        "description": "homework id",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.Homework"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update godoc",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "homework"
                ],
                "summary": "updates homework",
                "operationId": "homework-update",
                "parameters": [
                    {
                        "description": "updated homework",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.Homework"
                        }
                    },
                    {
                        "type": "string",
                        "description": "homework id",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.Homework"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete godoc",
                "tags": [
                    "homework"
                ],
                "summary": "deletes existing homework",
                "operationId": "homework-delete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "homework id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "data.Content": {
            "type": "object",
            "properties": {
                "Answer": {
                    "type": "string"
                },
                "Solution": {
                    "type": "string"
                },
                "Task": {
                    "type": "string"
                }
            }
        },
        "data.Data": {
            "type": "object",
            "properties": {
                "Homework": {
                    "$ref": "#/definitions/data.HomeworkData"
                }
            }
        },
        "data.Homework": {
            "type": "object",
            "properties": {
                "Id": {
                    "type": "integer"
                },
                "Name": {
                    "type": "string"
                }
            }
        },
        "data.HomeworkData": {
            "type": "object",
            "properties": {
                "Id": {
                    "type": "integer"
                },
                "Name": {
                    "type": "string"
                },
                "Questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/data.Question"
                    }
                }
            }
        },
...
        "data.Question": {
            "type": "object",
            "properties": {
                "Content": {
                    "$ref": "#/definitions/data.Content"
                },
                "Id": {
                    "type": "integer"
                },
                "Name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}